fix(subscription): No dispatch after dispose (#14)
__Issue:__

It's common for subscriptions in Onivim 2 to 'hold' the `dispatch`, like subscriptions that kick off an async operation, for example:

https://github.com/onivim/oni2/blob/3314264b65856b90640455c5573e93f946e00ed0/src/Service/Exthost/Service_Exthost.re#L617

However, the subscription might get disposed while the request is in-flight, and in that case, the `dispatch` function will be called, even though the subscription is not valid. This breaks the mental-model of subscriptions providing messages only when they are active.

As a work-around, in Onivim, there's a `Latch` that turns off dispatch after the subscription is disposed - but makes sense to just bake that in here. If there's a case where we are depending on a message to be dispatched _after_ a subscription is disposed, it's a logical bug.

__Fix:__ Add a ref that is carried along the lifecycle of the subscription - when disposed, we'll switch it to `false` so that future dispatches no longer occur.